Updated according to comments, renamed variables and changed logic of createTmpFile function within tests
