+++ cmd/keepalive/keepalive.go
@@ -41,7 +41,7 @@ func main() {
 		flag.Parse()
 	}
 
-	cfg, err := sf.GetConfigFromEnv([]sf.ConfigParam{
+	cfg, err := sf.GetConfigFromEnv([]*sf.ConfigParam{
 		{"Account", "SNOWFLAKE_TEST_ACCOUNT", true},
 		{"User", "SNOWFLAKE_TEST_USER", true},
 		{"Password", "SNOWFLAKE_TEST_PASSWORD", true},
